make script_bot compile in visual c++

+++ src/executor/unary_op.cpp
@@ -50,7 +50,7 @@ void unary_op(std::stack<object> &operands, op_code code) {
         } else {
             switch(code) {
             case op_code::plus:      return static_cast<std::int64_t>(v);
-            case op_code::negate:    return -v;
+			case op_code::negate:    return -static_cast<std::int64_t>(v);
             case op_code::bit_not:   return ~to_int(v);
             //case op_code::logic_not: return static_cast<std::int64_t>(!v);
             default:
